# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "Alert": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/AlertCode"
          },
          "column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column"
          },
          "evidence": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Evidence",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/AlertSeverity"
          }
        },
        "required": [
          "code",
          "severity",
          "message"
        ],
        "title": "Alert",
        "type": "object"
      },
...
      "ColumnProfile": {
        "properties": {
          "categorical": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoricalStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "dtype": {
            "title": "Dtype",
            "type": "string"
          },
          "n_unique": {
            "title": "N Unique",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "null_fraction": {
            "title": "Null Fraction",
            "type": "number"
          },
          "numeric": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NumericStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "semantic": {
            "$ref": "#/components/schemas/SemanticType"
          },
          "target_association": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "|Pearson| (numérica) o V de Cramér (categórica) con el target",
            "title": "Target Association"
          },
          "text": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TextStats"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "semantic",
          "dtype",
          "null_fraction",
          "n_unique"
        ],
        "title": "ColumnProfile",
        "type": "object"
      },
...
      "DataSource": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "additionalProperties": true,
            "description": "Sin secretos: usar `secret_refs` (keychain/vault)",
            "title": "Config",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inferred_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inferred Schema"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "secret_refs": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Secret Refs",
            "type": "object"
          },
          "type": {
            "$ref": "#/components/schemas/DataSourceType"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "default": 1,
            "description": "Versión para bloqueo optimista",
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "project_id",
          "name",
          "type"
        ],
        "title": "DataSource",
        "type": "object"
      },
      "DataSourceType": {
        "enum": [
          "file",
          "folder",
          "db",
          "hf",
          "kaggle",
          "api",
          "stream"
        ],
        "title": "DataSourceType",
        "type": "string"
      },
      "DatasetVersion": {
        "additionalProperties": false,
        "description": "Snapshot inmutable identificado por el hash de su manifiesto (RF-ING-07).",
        "properties": {
          "content_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "modality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Modality"
              },
              {
                "type": "null"
              }
            ]
          },
          "num_samples": {
            "minimum": 0.0,
            "title": "Num Samples",
            "type": "integer"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Directorio relativo al proyecto",
            "title": "Path"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "size_bytes": {
            "default": 0,
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "split": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Split"
              },
              {
                "type": "null"
              }
            ]
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "transformation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transformation"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "default": 1,
            "description": "Versión para bloqueo optimista",
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "project_id",
          "content_hash",
          "num_samples"
        ],
        "title": "DatasetVersion",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImageProfile": {
        "properties": {
          "channels": {
            "items": {
              "$ref": "#/components/schemas/CategoryCount"
            },
            "title": "Channels",
            "type": "array"
          },
          "corrupt": {
            "title": "Corrupt",
            "type": "integer"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "formats": {
            "items": {
              "$ref": "#/components/schemas/CategoryCount"
            },
            "title": "Formats",
            "type": "array"
          },
          "height_quantiles": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Height Quantiles",
            "type": "object"
          },
          "near_duplicate_fraction": {
            "title": "Near Duplicate Fraction",
            "type": "number"
          },
          "near_duplicate_pairs": {
            "title": "Near Duplicate Pairs",
            "type": "integer"
          },
          "resolutions": {
            "description": "Resoluciones WxH frecuentes (≥ k filas)",
            "items": {
              "$ref": "#/components/schemas/CategoryCount"
            },
            "title": "Resolutions",
            "type": "array"
          },
          "width_quantiles": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Width Quantiles",
            "type": "object"
          }
        },
        "required": [
          "count",
          "corrupt",
          "resolutions",
          "width_quantiles",
          "height_quantiles",
          "channels",
          "formats",
          "near_duplicate_pairs",
          "near_duplicate_fraction"
        ],
        "title": "ImageProfile",
        "type": "object"
      },
...
      "IngestBody": {
        "additionalProperties": false,
        "properties": {
          "overrides": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/SemanticType"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overrides"
          },
          "split": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SplitRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "title": "IngestBody",
        "type": "object"
      },
...
      "Modality": {
        "enum": [
          "tabular",
          "image",
          "text",
          "timeseries",
          "audio"
        ],
        "title": "Modality",
        "type": "string"
      },
...
      "ProfileCard": {
        "description": "Resumen agregado del dataset (solo train + val; el test sellado no se perfila).",
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/Alert"
            },
            "title": "Alerts",
            "type": "array"
          },
          "card_version": {
            "default": "1.0",
            "title": "Card Version",
            "type": "string"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnProfile"
            },
            "title": "Columns",
            "type": "array"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "dataset_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version Id"
          },
          "duplicate_row_fraction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Row Fraction"
          },
          "images": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "modality": {
            "$ref": "#/components/schemas/Modality"
          },
          "num_features": {
            "title": "Num Features",
            "type": "integer"
          },
          "num_samples": {
            "title": "Num Samples",
            "type": "integer"
          },
          "profiled_samples": {
            "title": "Profiled Samples",
            "type": "integer"
          },
          "split_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Split Counts",
            "type": "object"
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetProfile"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "modality",
          "num_samples",
          "profiled_samples",
          "split_counts",
          "num_features",
          "target",
          "columns"
        ],
        "title": "ProfileCard",
        "type": "object"
      },
...
      "SemanticType": {
        "enum": [
          "numeric",
          "categorical",
          "boolean",
          "datetime",
          "text",
          "id",
          "filepath"
        ],
        "title": "SemanticType",
        "type": "string"
      },
...
      "SourceCreate": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "path": {
            "description": "Archivo, carpeta o ZIP accesible por el Engine",
            "title": "Path",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/DataSourceType",
            "default": "file"
          }
        },
        "required": [
          "path"
        ],
        "title": "SourceCreate",
        "type": "object"
      },
...
      "SourcePreview": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "kind": {
            "$ref": "#/components/schemas/SourceKind"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "schema": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TableSchema"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "kind",
          "columns",
          "rows"
        ],
        "title": "SourcePreview",
        "type": "object"
      },
      "Split": {
        "additionalProperties": false,
        "properties": {
          "folds": {
            "anyOf": [
              {
                "minimum": 2.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folds"
          },
          "group_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Column"
          },
          "seed": {
            "default": 42,
            "title": "Seed",
            "type": "integer"
          },
          "strategy": {
            "$ref": "#/components/schemas/SplitStrategy"
          },
          "test": {
            "minimum": 0.0,
            "title": "Test",
            "type": "integer"
          },
          "test_sealed": {
            "default": true,
            "description": "El test no se usa en HPO (RF-ING-08)",
            "title": "Test Sealed",
            "type": "boolean"
          },
          "time_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Column"
          },
          "train": {
            "minimum": 0.0,
            "title": "Train",
            "type": "integer"
          },
          "val": {
            "minimum": 0.0,
            "title": "Val",
            "type": "integer"
          }
        },
        "required": [
          "strategy",
          "train",
          "val",
          "test"
        ],
        "title": "Split",
        "type": "object"
      },
...
      "TargetProfile": {
        "properties": {
          "classes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CategoryCount"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classes"
          },
          "imbalance_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "minoritaria / mayoritaria",
            "title": "Imbalance Ratio"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "numeric": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NumericStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "semantic": {
            "$ref": "#/components/schemas/SemanticType"
          },
          "task_hint": {
            "$ref": "#/components/schemas/TaskType"
          }
        },
        "required": [
          "name",
          "semantic",
          "task_hint"
        ],
        "title": "TargetProfile",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Perceptron Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/datasets/{dataset_version_id}": {
      "get": {
        "operationId": "getDataset",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_version_id",
            "required": true,
            "schema": {
              "title": "Dataset Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetVersion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dataset",
        "tags": [
          "data"
        ]
      }
    },
    "/api/v1/datasets/{dataset_version_id}/profile": {
      "get": {
        "operationId": "getProfile",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_version_id",
            "required": true,
            "schema": {
              "title": "Dataset Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileCard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile",
        "tags": [
          "data"
        ]
      },
      "post": {
        "operationId": "profileDataset",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_version_id",
            "required": true,
            "schema": {
              "title": "Dataset Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileCard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Profile",
        "tags": [
          "data"
        ]
      }
    },
    "/api/v1/datasets/{dataset_version_id}/samples": {
      "get": {
        "description": "Muestras para la UI (el test sellado no se expone).",
        "operationId": "getDatasetSamples",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_version_id",
            "required": true,
            "schema": {
              "title": "Dataset Version Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "split",
            "required": false,
            "schema": {
              "default": "train",
              "title": "Split",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Getdatasetsamples",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dataset Samples",
        "tags": [
          "data"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/datasets": {
      "get": {
        "operationId": "listDatasets",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetVersion"
                  },
                  "title": "Response Listdatasets",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Datasets",
        "tags": [
          "data"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/sources": {
      "post": {
        "operationId": "createSource",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Source",
        "tags": [
          "data"
        ]
      }
    },
...
    "/api/v1/sources/{source_id}/ingest": {
      "post": {
        "operationId": "ingestSource",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetVersion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Source",
        "tags": [
          "data"
        ]
      }
    },
    "/api/v1/sources/{source_id}/preview": {
      "post": {
        "operationId": "previewSource",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcePreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Source",
        "tags": [
          "data"
        ]
      }
    },
...
  }
}